Implement mergesort to be used in `array_merge.c`
We had some issues with includes of the mergesort files, as
well as the syntax of includes that worked fine in the context of
just mergesort, but the compiler would complain when those includes
were invoked in `array_merge`.

We eventually figured out how to get the relative path, and for some
reason when relatively invoking mergesort we had to also include its
`.c` version too.