Resolve our issue debugging array_merge in step 1
We were using main in `array_merge.c`'s main, which we only have
a main for for easy debugging. We were just seeing if we could merge
all the values. Instead of our sytax that was failing, we decided to
make the sysntax of example data like the .cpp tests, and it was
valid! Not arguing.

As of this commit, we print out the presorted values, the amount of unique
values, and the sorted array.